Reorganized the project folders Created makefile Added gitignore file
