+++ .gitignore
@@ -1,0 +1,4 @@
+build
+*.swp
+*.o
+*.d
